requirements.txt
# PostgreSQL adapter - Using psycopg instead of psycopg2 for Python 3.13 compatibility
psycopg[binary]>=3.1.12
pytest>=7.4.0
python-dotenv>=1.0.0